Move over utils. Add docs on populate.
